// src/schemas.g.ts
// Generated by ts-to-zod
import { z } from "zod";

export const jobRequestSchema = z.object({
  yarn: z.string(),
  author: z.string().optional(),
  title: z.string().optional(),
  debug: z.boolean().optional(),
  shuffle: z.boolean().optional(),
});

export const jobStateSchema = z.union([
  z.literal("Unknown"),
  z.literal("Failed"),
  z.literal("Processing"),
  z.literal("Complete"),
]);

export const pDFGenerationReponseSchema = z.object({
  jobID: z.string(),
  state: jobStateSchema,
  message: z.string().optional().nullable(),
  details: z.string().optional().nullable(),
  pdfLocation: z.string().optional().nullable(),
});